final touch ups for version 3.3

+++ src-tauri/src/constants.rs
@@ -78,4 +78,4 @@ pub fn get_local_versions_path() -> String {
 }
 
 pub const MAJOR_VERSION: f32 = 3.3;
-pub const MINOR_VERSION: u8 = 0;
+pub const MINOR_VERSION: i64 = 0;
